Added Update Stock Client
1. Stock Update Service added
2.Console Client To Update api call added

+++ ChannelEngine/ChannelEngine.BusinessLogic/Models/StockLocations.cs
@@ -8,6 +8,6 @@ namespace ChannelEngine.Business.Models
     public class StockLocations
     {
         public long Stock { get; set; }
-        public long StockLocationId { get; set; }
+        public int StockLocationId { get; set; }
     }
 }
